I am sick and tired of 32 bit integers

+++ src/template.kt
@@ -1,16 +1,16 @@
 fun main() {
-    fun part1(input: List<String>): Int {
-        return -1
+    fun part1(input: List<String>): Long {
+        return -1L
     }
 
-//    fun part2(input: List<String>): Int {
+//    fun part2(input: List<String>): Long {
 //
 //    }
 
     val testInput = readInput("day{day}_test")
 
-    check(part1(testInput) == 0)
-//    check(part2(testInput) == 0)
+    check(part1(testInput) == 0L)
+//    check(part2(testInput) == 0L)
 
     val input = readInput("day{day}")
     println(part1(input))
